requirements.txt
Django==3.1.6
django-heroku==0.3.1
django-import-export==2.5.0
gunicorn==20.0.4
mysqlclient==2.0.3
